Refactor SidebarNav
