+++ src/components/controllers/page/sidebar.js
@@ -1,15 +1,17 @@
 import { h } from 'preact';
 import style from './sidebar.less';
 import DocVersion, { getCurrentDocVersion } from './../../doc-version';
-import Toc from './table-of-content';
+import SidebarNav from './sidebar-nav';
 import { useState, useCallback } from 'preact/hooks';
 import config from '../../../config.json';
 
-export default function Sidebar(props) {
+export default function Sidebar() {
 	const [open, setOpen] = useState(false);
 	const toggle = useCallback(() => setOpen(!open), [open]);
 	const close = useCallback(() => setOpen(false));
 	const version = getCurrentDocVersion();
+
+	// Get menu items for the current version of the docs
 	const docNav = config.docs
 		.filter(item => item.path.indexOf(`/v${version}`) > -1)
 		.map(item => ({ text: item.name, level: 2, href: item.path }));
@@ -22,8 +24,7 @@ export default function Sidebar(props) {
 			<aside class={style.sidebar}>
 				<div class={style.sidebarInner}>
 					<DocVersion />
-					{/* <Toc items={props.toc} onClick={close} /> */}
-					<Toc items={docNav} onClick={close} />
+					<SidebarNav items={docNav} onClick={close} />
 				</div>
 			</aside>
 		</div>
